% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cleanStringVec2013}
\alias{cleanStringVec2013}
\title{Clean name strings}
\usage{
cleanStringVec2013(stringVec)
}
\arguments{
\item{stringVec}{character}
}
\value{
character
}
\description{
Clean name strings
}
