% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cleanStringVec2018}
\alias{cleanStringVec2018}
\title{Clean name strings}
\usage{
cleanStringVec2018(stringVec)
}
\arguments{
\item{stringVec}{character}
}
\value{
character
}
\description{
Clean name strings
}
