% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make2013CensusTibble}
\alias{make2013CensusTibble}
\title{Create tibble from the census csv frames}
\usage{
make2013CensusTibble(dflist, agg_lv = "mb", sp_unit_ids = NA)
}
\arguments{
\item{dflist}{list}

\item{agg_lv}{character}

\item{sp_unit_ids}{character}
}
\value{
tibble
}
\description{
Create tibble from the census csv frames
}
