% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make2018CensusTibble}
\alias{make2018CensusTibble}
\title{Create tibble from the census csv frames}
\usage{
make2018CensusTibble(dflist, looklist, sp_unit_ids = NA, keep_2018 = F)
}
\arguments{
\item{dflist}{list}

\item{looklist}{list}

\item{sp_unit_ids}{character}

\item{keep_2018}{logical}
}
\value{
tibble
}
\description{
Create tibble from the census csv frames
}
