% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeCensusTibble2018}
\alias{makeCensusTibble2018}
\title{Create 'full' census tibble from dflist (from makeDflist)}
\usage{
makeCensusTibble2018(dflist)
}
\arguments{
\item{dflist}{list}
}
\value{
tibble
}
\description{
Create 'full' census tibble from dflist (from makeDflist)
}
