% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeDflist2013}
\alias{makeDflist2013}
\title{Create dflist of the census csv frames}
\usage{
makeDflist2013(path, files, mbs)
}
\arguments{
\item{path}{string}

\item{files}{character}

\item{mbs}{character}
}
\value{
list
}
\description{
Create dflist of the census csv frames
}
