% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeDflist2018}
\alias{makeDflist2018}
\title{Create dflist of the census csv frames}
\usage{
makeDflist2018(path, files, lookups, sas)
}
\arguments{
\item{path}{character}

\item{files}{character}

\item{lookups}{character}

\item{mbs}{character}
}
\value{
list
}
\description{
Create dflist of the census csv frames
}
