% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opdorg-package.R
\docType{package}
\name{opdorg-package}
\alias{opdorg}
\alias{opdorg-package}
\title{opdorg: opdorg - Open Data Organiser}
\description{
Gets, cleans, and tidies various NZ open source data sets. Currently implements functions to access and tidy NZ census meshblock and sa1 datasets between 2001 and 2018.
}
\author{
\strong{Maintainer}: Leon Stirk-Wang \email{leon.stirkwang@protonmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
