% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveWFS.R
\name{retrieveWFS}
\alias{retrieveWFS}
\title{Retrieve WFS layer data in WGS84 coordinate format}
\usage{
retrieveWFS(url, bounding_data, layer, crs)
}
\arguments{
\item{url}{character}

\item{bounding_data}{data.frame}

\item{layer}{character}

\item{crs}{integer}
}
\value{
sf
}
\description{
Retrieve WFS layer data in WGS84 coordinate format
}
