% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveREST.R
\name{retrieveREST}
\alias{retrieveREST}
\title{Retrieve REST layer data in WGS84 coordinate format}
\usage{
retrieveREST(url, feature, layer, id_var_name)
}
\arguments{
\item{url}{character}

\item{feature}{character}

\item{layer}{character}

\item{id_var_name}{character}
}
\value{
sf
}
\description{
Retrieve REST layer data in WGS84 coordinate format
}
